muevo generate_multiple_images a GenrationGI0.py
