+++ notebooks/scripts/main.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# In[2]:
+# In[ ]:
 
 
 import sys
 sys.path.append('..')
 
-from scripts.GenrationGI0 import partitioned_gi0_image
-from scripts.autoencoders import InMemoryImageDataset, generate_multiple_images, ConfigurableAutoencoder
+from scripts.GenrationGI0 import partitioned_gi0_image, generate_multiple_images
+from scripts.autoencoders import InMemoryImageDataset, ConfigurableAutoencoder
 
 import pandas as pd
 import torch
