+++ notebooks/scripts/visualizar_imagenes.py
@@ -1,14 +1,13 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# In[1]:
+# In[ ]:
 
 
 import sys
 sys.path.append('..')
 
-from scripts.GenrationGI0 import partitioned_gi0_image
-from scripts.autoencoders import generate_multiple_images
+from scripts.GenrationGI0 import partitioned_gi0_image, generate_multiple_images
 
 import numpy as np
 import matplotlib.pyplot as plt
